another changes

+++ backend/routes/review.js
@@ -3,6 +3,24 @@ const router = express.Router()
 const pool = require('../utils/db')
 const result = require('../utils/result')
 
+
+router.get('/display', (req, res) => {
+    const sql = "SELECT * FROM reviews"
+    pool.query(sql, (error, data) => {
+        console.log(error)
+        res.send(result.createResult(error, data))
+    })
+})
+
+router.get('/display/my', (req, res) => {
+    const { user_id } = req.body
+    const sql = "SELECT * FROM reviews WHERE user_id = ?"
+    pool.query(sql, [user_id], (error, data) => {
+        console.log(error)
+        res.send(result.createResult(error, data))
+    })
+})
+
 router.post('/create', (req, res) => {
     const { movie_id, review, rating, user_id, modified } = req.body
     const sql = "INSERT INTO reviews(movie_id, review, rating, user_id, modified) VALUES(?, ?, ?, ?, ?)"
@@ -22,6 +40,14 @@ router.put('/edit', (req, res) => {
 })
 
 router.delete('/delete', (req, res) => {
-    const {movie_id, user_id} = req.body
-    const sql = "DELETE FROM "
+    const { movie_id, user_id } = req.body
+    const sql = "DELETE FROM reviews WHERE movie_id = ? AND user_id = ?"
+    pool.query(sql, [movie_id, user_id], (error, data) => {
+        console.log(error)
+        res.send(result.createResult(error, data))
+    })
 })
+
+
+
+module.exports = router
